+++ proto/buf.yaml
@@ -5,6 +5,7 @@ deps:
   - buf.build/googleapis/googleapis
   - buf.build/bufbuild/protovalidate
   - buf.build/pentops/o5
+  - buf.build/pentops/j5
 
 lint:
   use:
