+++ proto/o5/registry/builder/v1/builder.proto
@@ -1,0 +1,41 @@
+syntax = "proto3";
+
+package o5.registry.builder.v1;
+
+import "buf/validate/validate.proto";
+import "google/protobuf/empty.proto";
+import "google/protobuf/timestamp.proto";
+import "j5/config/v1/config.proto";
+import "j5/source/v1/image.proto";
+import "o5/messaging/v1/annotations.proto";
+
+option go_package = "github.com/pentops/registry/gen/o5/registry/builder/v1/builder_tpb";
+
+service BuilderTopic {
+  option (o5.messaging.v1.config).unicast.name = "j5-builder-input";
+  rpc BuildProto(BuildProtoMessage) returns (google.protobuf.Empty) {}
+  rpc BuildAPI(BuildAPIMessage) returns (google.protobuf.Empty) {}
+}
+
+message CheckRun {
+  string name = 1;
+  int64 id = 2;
+}
+
+message BuildProtoMessage {
+  j5.source.v1.CommitInfo commit = 1 [(buf.validate.field).required = true];
+  CheckRun check_run = 2 [(buf.validate.field).required = true];
+
+  // contains one single proto build definition, otherwise is the config
+  // as pulled from the source.
+  j5.config.v1.Config config = 3 [(buf.validate.field).required = true];
+}
+
+message BuildAPIMessage {
+  j5.source.v1.CommitInfo commit = 1 [(buf.validate.field).required = true];
+  CheckRun check_run = 2 [(buf.validate.field).required = true];
+
+  // contains fields relating to JSON API build, i.e. excludes other build
+  // configs.
+  j5.config.v1.Config config = 3 [(buf.validate.field).required = true];
+}
